+++ benchmarks/uvgrtp/sender.cc
@@ -16,7 +16,7 @@ void thread_func(void *mem, size_t len, char *addr, int thread_num, double fps, 
     uint64_t total_size = 0;
     uint64_t diff       = 0;
 	uint64_t current    = 0;
-	uint64_t period     = (uint64_t)((1000 / fps) * 1000);
+	uint64_t period     = (uint64_t)((1000 / (float)fps) * 1000);
     rtp_error_t ret     = RTP_OK;
     std::string addr_("10.21.25.2");
